user and state complected wit some changes in the table master_state -> is deleted column -> short to int.

+++ tmsBackEnd/src/main/java/com/yash/tms/services/MasterStateManager.java
@@ -6,7 +6,7 @@ import com.yash.tms.entity.MasterState;
 
 public interface MasterStateManager {
 
-	public List<MasterState> findAllStates(short stateIsDeleted);
+	public List<MasterState> findAllStates(int stateIsDeleted);
 
 	public MasterState addState(MasterState state);
